refactor(channel): extract View() into channel_view.go
Phase 2 continuation. The 278-LOC Bubbletea View() method moves out of
channel.go into a kin file. View is the render entrypoint — composes
sidebar | main | thread, builds the header strip + status bar, and
joins the panels via lipgloss.JoinHorizontal. The presentation
primitives (channelui.RenderXxx, MainPanelStyle, etc.) live in the
channelui package; View() is the orchestrator.

  - cmd/wuphf/channel.go: -279 LOC (2579 -> 2300)
  - cmd/wuphf/channel_view.go: +298 LOC (new)

Cumulative channel.go reduction across Phase 2:
  4807 -> 2300
  Net: -2507 LOC; channel.go shrunk by 52%.

No behavior change. cmd/wuphf tests pass.
